criando CalculoRendaFixaUtil

+++ src/main/java/br/com/smartinvest/smart_invest_api/model/Simulacao.java
@@ -43,7 +43,7 @@ public class Simulacao {
     private Integer prazoMeses;
 
     @Column(name = "taxaJuros")
-    private Double taxaJuros;
+    private BigDecimal taxaJuros;
 
     @Column(name = "valorFinal")
     private BigDecimal valorFinal;
